+++ tests/testthat/test-data_arrange.R
@@ -1,4 +1,4 @@
-library(dplyr)
+library(poorman)
 df <- head(mtcars)
 df$character <- c("a", "b", "b", "c", "c", "a")
 
@@ -44,10 +44,17 @@ test_that("data_arrange works without columns", {
 })
 
 test_that("data_arrange ignores wrong names if safe = TRUE", {
-  expect_equal(data_arrange(df, "foo"), df)
-  expect_equal(
-    data_arrange(df, c("gear", "foo")),
-    data_arrange(df, "gear")
+  expect_warning(
+    expect_equal(data_arrange(df, "foo"), df),
+    regexp = "don't exist"
+  )
+
+  expect_warning(
+    expect_equal(
+      data_arrange(df, c("gear", "foo")),
+      data_arrange(df, "gear")
+    ),
+    regexp = "don't exist"
   )
 })
 
@@ -62,4 +69,62 @@ test_that("data_arrange errors if not coercable to data frame", {
     structure(list(a = 5:1, b = c("a", "b", "c", "d", "e")), row.names = 5:1, class = "data.frame"),
     ignore_attr = TRUE
   )
+})
+
+test_that("data_arrange works with grouped df", {
+  set.seed(123)
+  x <- mtcars[sample(1:nrow(mtcars), 10, replace = TRUE), c("cyl", "mpg")]
+  g <- data_group(x, cyl)
+
+  expected <- data.frame(
+    cyl = c(4, 4, 4, 6, 6, 8, 8, 8, 8, 8),
+    mpg = c(22.8, 30.4, 32.4, 17.8, 19.2, 10.4, 15, 15.2, 15.5, 18.7)
+  )
+  class(expected) <- c("grouped_df", "data.frame")
+  rownames(expected) <- c(
+    "Datsun 710", "Honda Civic", "Fiat 128", "Merc 280C", "Merc 280",
+    "Cadillac Fleetwood", "Maserati Bora", "Merc 450SLC", "Dodge Challenger",
+    "Hornet Sportabout"
+  )
+  attributes(expected)$groups <- attributes(g)$groups
+
+  expect_equal(
+    data_arrange(g, "mpg"),
+    expected
+  )
+})
+
+test_that("data_arrange works with NA", {
+
+  # without groups
+
+  tmp <- data.frame(
+    a = c(1, 2, 2, 8, 1, 3),
+    b = c(1, NA, 3, 3, NA, 5)
+  )
+
+  expect_equal(
+    data_arrange(tmp, "a"),
+    data.frame(
+      a = c(1, 1, 2, 2, 3, 8),
+      b = c(1, NA, NA, 3, 5, 3)
+    )
+  )
+
+  # with groups
+
+  g <- data_group(tmp, "b")
+
+  expected <- data.frame(
+    a = c(1, 2, 8, 3, 1, 2),
+    b = c(1, 3, 3, 5, NA, NA)
+  )
+  class(expected) <- c("grouped_df", "data.frame")
+  attributes(expected)$groups <- attributes(g)$groups
+
+  expect_equal(
+    data_arrange(g, "a"),
+    expected
+  )
+
 })